Merges dev into main with MVP.

+++ RSGym_Client/Actions/DeleteRequestAction.cs
@@ -3,11 +3,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace RSGym_Client
 {
-    class DeleteRequestAction : IBaseAction
+    class DeleteRequestAction : IBaseAction, ICommunicable
     {
 
         #region Properties
@@ -22,6 +21,8 @@ namespace RSGym_Client
 
         public bool Success { get; set; }
 
+        public string FeedbackMessage { get; set; }
+
         #endregion
 
         #region Contructor
@@ -32,6 +33,8 @@ namespace RSGym_Client
             Name = "Delete request";
             User = new GuestUser();
             MenuType = MenuType.Restricted;
+            Success = false;
+            FeedbackMessage = string.Empty;
         }
 
         #endregion
@@ -41,38 +44,64 @@ namespace RSGym_Client
         public void Execute(out bool isExit)
         {
             isExit = false;
+            string deletedRequest = string.Empty;
 
-            // 1. Listar apenas pedidos agendados do user
-            List<Request> scheduledRequests = RequestRepository.GetRequestsByUserID(this.User.UserID)
-                .Where(r => r.Status == RequestStatus.Agendado).ToList();
+            List<Request> scheduledRequests = RequestRepository
+                .GetRequestsByUserID(this.User.UserID)
+                .Where(r => r.Status == RequestStatus.Agendado)
+                .ToList();
 
             Console.WriteLine("\nEscolha um pedido para cancelar/apagar");
 
-            string requestHeader = scheduledRequests.GetHeader(out int trainerLength, out int statusLength, out int messageLength);
+            string requestHeader = scheduledRequests
+                .GetRequestHeader(out int trainerLength, out int statusLength, out int messageLength);
 
             Console.WriteLine(requestHeader);
-            scheduledRequests.ForEach(r => Console.WriteLine(r.ToString(trainerLength, statusLength, messageLength)));
+            scheduledRequests
+                .ForEach(r => Console.WriteLine(r.ToString(trainerLength, statusLength, messageLength)));
 
-            // 2. Selecionar um dos pedidos, pelo id
             Console.Write("\nOpção selecionada: ");
             string request = this.ReadUserInput();
 
-            // toDo: Validate if input is integer
-            int requestID = int.Parse(request);
+            _ = int.TryParse(request, out int requestID);
 
-            Request currentRequest = scheduledRequests.Where(r => r.RequestID == requestID).Single();
+            Request currentRequest = scheduledRequests
+                .Where(r => r.RequestID == requestID)
+                .FirstOrDefault();
+            
 
-            RequestRepository.DeleteRequestByID(requestID);
-            Success = true;
+            if (currentRequest != null)
+            {
+                deletedRequest = $"{requestHeader}\n{currentRequest.ToString(trainerLength, statusLength, messageLength)}";
+                RequestRepository.DeleteRequestByID(requestID);
+            }
+
+            Success = !(currentRequest is null);
+
+            BuildFeedbackMessage(deletedRequest, requestID);
 
             Console.Clear();
+        }
 
+        public void BuildFeedbackMessage(string request = "", int requestID = 0)
+        {
             var sb = new StringBuilder();
-            sb.AppendLine("Pedido cancelado/apagado:");
-            sb.AppendLine(requestHeader);
-            sb.Append(currentRequest.ToString(trainerLength, statusLength, messageLength));
 
-            Communicator.WriteSuccessMessage(sb.ToString());
+            if (Success)
+            {
+                sb.AppendLine("Pedido cancelado/apagado:");
+                sb.Append(request);
+            }
+            else if(requestID == 0)
+            {
+                sb.Append("Selecione um pedido válido.");
+            }
+            else
+            {
+                sb.Append($"Não foi localizado um pedido com o nº {requestID} nas sua lista de pedidos.");
+            }
+
+            FeedbackMessage = sb.ToString();
         }
 
         #endregion
